Fixes and Code reorganization and Optimization
