+++ backend/package.json
@@ -27,11 +27,11 @@
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.14",
-    "@types/node": "^22.9.0",
+    "@types/node": "^22.10.9",
     "@types/passport-google-oauth20": "^2.0.16",
     "@types/supertest": "^6.0.2",
     "concurrently": "^9.1.0",
-    "express": "^4.21.1",
+    "express": "^4.21.2",
     "jest": "^29.7.0",
     "nodemon": "^3.1.7",
     "supertest": "^7.0.0",
@@ -43,6 +43,7 @@
     "@types/express-session": "^1.18.1",
     "@types/passport": "^1.0.17",
     "bcrypt": "^5.1.1",
+    "body-parser": "^1.20.3",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
@@ -52,6 +53,7 @@
     "passport": "^0.7.0",
     "passport-google-oauth20": "^2.0.0",
     "pg": "^8.13.1",
-    "sequelize": "^6.37.5"
+    "sequelize": "^6.37.5",
+    "xendit-node": "^6.0.0"
   }
 }
